requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
clipboard==0.0.4
idna==3.6
lxml==5.1.0
pillow==10.2.0
pyperclip==1.8.2
requests==2.31.0
soupsieve==2.5
urllib3==2.2.1
